Cleanup
Fixed bug where combo finisher's explosion didn't come from site of projectile impact, combo finishers' bullet speeds now vary depending on the combo, Attack infos can now be added together, UI updated to fit my computer

+++ Assets/Scripts/Projectile Scripts/QuarterNoteProjectile.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class QuarterNoteProjectile : ProjectileBase
 {
 
-    public override void setCourseOfFire(int bulletSpeed, bool facingRight, Vector3 shootPos, HashSet<string> targetTags)
+    public override void setCourseOfFire(float bulletSpeed, bool facingRight, Vector3 shootPos, HashSet<string> targetTags)
     {
         projTargetTags = targetTags;
         Rigidbody2D rb = this.GetComponent<Rigidbody2D>();
